mcp/mcp-server/requirements.txt
psycopg2
fastapi
uvicorn
load_dotenv
azure.mgmt.loganalytics
azure.identity
requests